Have most of the proposed changes working, separating JADE from MUSCLE.
One important problem is introduced: the takedown of instancecontrollers
is not completely registered and seems to interfere with messages being
sent. In SimpleExample.cxa.rb, four messages are sent, the Sender(w) is
taken down but the last message is not processed by ConsoleWriter.

Other problem: filters are now applied in the conduitentrancecontroller,
but this unfortunately means these filters act on the pointer to the
data that the writer is using, modifying the original data. This does save
memory but may have unwanted side effects in the sending code. In SimpleExample
it means that the array that is sent is divided by two each iteration.

git-svn-id: https://apps.man.poznan.pl/svn/muscle/muscle/branches/filter@127 b725f6ce-2418-43e8-b868-1e0ad3d2b257