muscle::env::init now initializes a new muscle instance if its port is not given as an environment variable.
This required minor change in the signature: from "void muscle::env::init(void)" to "int muscle::env::init(int *argc, char ***argv)". When an executable is compiled, it can be invoked with "./EXE args -- muscle_args". In the case of LHC, this could be
"./LHC -- --main --cxa LHC2.cxa.rb LHC" The advantage of this setup is that you can now prepend the command with "mpiexec -np n", or run it from qsub, and everything behaves as it should.

The host address and port of Java MUSCLE are written to a temporary fifo that is initialized by env::init, written to by MUSCLE and then read from by env::init.

The next step would be to initialize Java MUSCLE on another host, let it write its port to a file, and run the C++ client somewhere else, giving it the environment variables to contact MUSCLE.

git-svn-id: https://apps.man.poznan.pl/svn/muscle/muscle/trunk@396 b725f6ce-2418-43e8-b868-1e0ad3d2b257