[BUILD] Enhanced build script so also users can use it
It now takes an argument INSTALL_PREFIX, which overrides the config file
and will install in any given directory. If there is no config file
present for the host, it will only install a release version, and not in
the devel / devel-debug directories. This would be the default behaviour
that users encounter.


git-svn-id: https://apps.man.poznan.pl/svn/muscle/muscle/trunk@571 b725f6ce-2418-43e8-b868-1e0ad3d2b257