[Java] In the tmp dir make symlinks from manager to simulations and vice-versa
This way it's easier to group files that belong to a single simulation.
It is implemented by adding a unique directory name to TcpLocation, which will be used
to create the symlink from during registration. It also requires one additional RTT
when a muscle instance is started, to get the manager location.



git-svn-id: https://apps.man.poznan.pl/svn/muscle/muscle/trunk@687 b725f6ce-2418-43e8-b868-1e0ad3d2b257