[Java/C++] Not copying data coming from native conduits
This is achieved by a new addSharedDataEntrance call in the API (besides also a addAsynchronousEntrance).
In this shared data entrance only a pointer is copied to other submodels, not the data itself. It is not
the default, since filters and the other submodels can modify the original data. The addAsynchronousEntrance
copies the data as soon as it is sent, but the transmission is done while the submodel continues to compute.
It is not copied again before transmission.



git-svn-id: https://apps.man.poznan.pl/svn/muscle/muscle/trunk@713 b725f6ce-2418-43e8-b868-1e0ad3d2b257