[Java] Changes to have MUSCLE working reliably with up to 1000 submodels, 50 edges each
For this, there is a new Thread pool that does not construct new Threads immediately, but only if all other threads are blocked, used for the PortFactory
To prevent overload of the socket of the simulationmanager, causing Socket reset, the TcpIDManipulator now uses a socketpool, reusing old sockets if possible. The simulation manager recognizes this as a separate protocol, and can also communicate with the old protocol.
The SimulationManager, TCPIDManipulator, DelegatingResolver and InstanceID now have much less synchrnonization blocks, and DelegatingResolver now locks per ID and doesn't lock the whole class


git-svn-id: https://apps.man.poznan.pl/svn/muscle/muscle/trunk@742 b725f6ce-2418-43e8-b868-1e0ad3d2b257