[Java] Added the possibility to run submodels and mappers asynchronously
In the client, its possible to specify the number of runner threads (-t). For 2000 submodels, you can use 2 threads which
will step through all submodels and execute them as soon as they are ready. The MultiControllerRunner is used for this.
If an instance is not ready for stepping, a separate thread is spawned anyway.
In the mapper, there are now readyForReceive, readyForSend, and readyForContinue
In the submodel, there are readyForInit, readyForIntermediateObservation, readyForFinalObservation, readyForSolvingStep
These should only return true if they can complete without blocking.



git-svn-id: https://apps.man.poznan.pl/svn/muscle/muscle/trunk@754 b725f6ce-2418-43e8-b868-1e0ad3d2b257